난이도에 따른 속도 조정 - 이현준

+++ shoottheduck/src/main/java/kr/jbnu/se/std/BadDuck.java
@@ -3,7 +3,7 @@ package kr.jbnu.se.std;
 import java.awt.image.BufferedImage;
 
 public class BadDuck extends Duck {
-    public BadDuck(int x, int y, int speedX, int speedY, BufferedImage badDuckImg) {
+    public BadDuck(int x, int y, double speedX, int speedY, BufferedImage badDuckImg) {
         super(x, y, speedX, speedY, badDuckImg);
         this.score = -50; // 플레이어가 쏘면 -50점을 잃음
     }
